% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{busco_tree_parsed}
\alias{busco_tree_parsed}
\title{Get BUSCO phylogeny}
\usage{
busco_tree_parsed(paths, rename = TRUE)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}

\item{rename}{If \code{TRUE}, rename the tip labels to sample IDs and reference IDs.}
}
\value{
List of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path
}
\description{
Return a list of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path by searching for
folders containing pathogensurveillance output.
}
\seealso{
Other parsers: 
\code{\link{core_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
