% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{clean_input_table}
\alias{clean_input_table}
\title{Fix formatting errors}
\usage{
clean_input_table(table, known_columns)
}
\arguments{
\item{table}{The input table}

\item{known_columns}{Column names to convert similar column names to}
}
\description{
Removes empty columns and rows, cleans up known column names, remove
duplicate columns, trims excess whitespace, etc
}
\keyword{internal}
