% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{core_ref_path}
\alias{core_ref_path}
\title{Find the core gene analysis reference data file path}
\usage{
core_ref_path(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain pathogensurveillance output.}
}
\value{
character vector of length 1
}
\description{
Return the file path to the TSV with the list of references used in the core
gene analysis for a given pathogensurveillance output folder.
}
\seealso{
Other path finders: 
\code{\link{busco_ref_path}()},
\code{\link{busco_tree_path}()},
\code{\link{considered_ref_meta_path}()},
\code{\link{core_tree_path}()},
\code{\link{estimated_ani_matrix_path}()},
\code{\link{pocp_matrix_path}()},
\code{\link{ref_meta_path}()},
\code{\link{run_info_path}()},
\code{\link{sample_meta_path}()},
\code{\link{selected_ref_meta_path}()},
\code{\link{sendsketch_path}()},
\code{\link{software_version_path}()},
\code{\link{status_message_path}()},
\code{\link{variant_align_path}()},
\code{\link{variant_ref_path}()},
\code{\link{variant_tree_path}()}
}
\concept{path finders}
