% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{core_tree_plot}
\alias{core_tree_plot}
\title{Plot core gene phylogeny}
\usage{
core_tree_plot(input, collapse_by_tax = NULL, interactive = FALSE)
}
\arguments{
\item{input}{The path to one or more folders that contain
pathogensurveillance output or paths to tree files.}

\item{collapse_by_tax}{A \code{\link[base:character]{base::character()}} vector of taxonomic
classifications, each delimited by \verb{;}, and named by sample or reference
ids present in \code{sample_meta} or \code{ref_meta}. These are used to provide a
taxonomic tree that functions as a backbone to combine multiple trees into
a single one. (Default: return a list of plots instead of a single plot)}

\item{interactive}{Whether to use an HTML-based interactive format or not
(default: TRUE)}
}
\value{
A list of plots, unless \code{collapse_by_tax} is used, in which case a single plot is returned.
}
\description{
Plot the core gene phylogenies present in the output of a pathogensurveillance
run.
}
