% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{default_column_values}
\alias{default_column_values}
\title{Defaults for known columns}
\usage{
default_column_values()
}
\value{
A \code{vector} of default values named by column names.
}
\description{
These are used if a cell is left blank.
}
\examples{
default_column_values()

}
