% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{estimated_ani_match_table}
\alias{estimated_ani_match_table}
\title{Print best ANI match table}
\usage{
estimated_ani_match_table(path, interactive = FALSE, ...)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{interactive}{Whether to produce interactive tables (TRUE) or static
tables (FALSE). Defaults to TRUE if the environment supports HTML output,
otherwise FALSE. Interactive tables offer enhanced browsing capabilities,
while static tables are best for printed pdf reports.}

\item{...}{Passed to \code{DT::datatable} for interactive output.}
}
\description{
Prints a table with the highest ANI matches for each sample given a pairwise
distance matrix.
}
