% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{estimated_ani_matrix_parsed}
\alias{estimated_ani_matrix_parsed}
\title{Get estimated ANI distance matrix}
\usage{
estimated_ani_matrix_parsed(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
a \code{list} of ANI matrices
}
\description{
Return a list of \code{\link[base:data.frame]{base::data.frame()}}s with the estimated pairwise ANI values
calculated by sourmash.
}
\examples{
path <- system.file('extdata/ps_output', package = 'psminer')
estimated_ani_matrix_parsed(path)

}
