% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{find_group_result_path}
\alias{find_group_result_path}
\title{Find subfolders with report group results}
\usage{
find_group_result_path(paths)
}
\description{
For one or more paths, return the paths of subfolders (or the given folders)
that are the report group output of a pathogensurviellance run.
}
\keyword{internal}
