% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{find_path_report_group}
\alias{find_path_report_group}
\title{Determine the report group for files}
\usage{
find_path_report_group(path)
}
\description{
Determine the report group for files
}
\keyword{internal}
