% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_number}
\alias{format_number}
\title{Format a number for printing}
\usage{
format_number(nums, sig_fig = 4)
}
\description{
Shortens a number to a specified number of significant digits.
Taken from https://stackoverflow.com/questions/3245862/format-numbers-to-significant-figures-nicely-in-r
}
\keyword{internal}
