% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{generalized_tree_plot}
\alias{generalized_tree_plot}
\title{Plot generic phylogeny}
\usage{
generalized_tree_plot(
  path,
  parser,
  collapse_by_tax = NULL,
  interactive = FALSE
)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output or paths to tree files.}

\item{parser}{A function that takes directory paths as input and returns
parsed trees found in those directories, such as
\code{\link[=core_tree_parsed]{core_tree_parsed()}}.}

\item{collapse_by_tax}{A \code{\link[base:character]{base::character()}} vector of taxonomic
classifications, each delimited by \verb{;}, and named by sample or reference
ids present in \code{sample_meta} or \code{ref_meta}. These are used to provide a
taxonomic tree that functions as a backbone to combine multiple trees into
a single one. (Default: return a list of plots instead of a single plot)}

\item{interactive}{Whether to use an HTML-based interactive format or not
(default: TRUE)}
}
\value{
A list of plots, unless \code{collapse_by_tax} is used, in which case a
single plot is returned.
}
\description{
Plot phylogenies present in the output of a pathogensurveillance run.
}
\keyword{internal}
