% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDOI.R
\name{get_bibtex_from_doi}
\alias{get_bibtex_from_doi}
\title{Get BibTeX Citation from DOI}
\usage{
get_bibtex_from_doi(doi, ...)
}
\arguments{
\item{doi}{The DOI (Digital Object Identifier) for the publication.}
}
\value{
A character string containing the BibTeX citation.
}
\description{
This function retrieves BibTeX citation from a given DOI using the rcrossref package.
}
\details{
R Package Header Example
}
\examples{
\dontrun{
doi <- "https://doi.org/10.1002/cpbi.102"
bibtex_citation <- get_bibtex_from_doi(doi)
cat(bibtex_citation)
}

}
\author{
Ricardo I. Alcala
}
\keyword{BibTeX}
\keyword{DOI}
\keyword{rcrossref}
