% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{get_ncbi_assemblies}
\alias{get_ncbi_assemblies}
\title{Get metadata for NCBI assemblies using a query}
\usage{
get_ncbi_assemblies(
  query,
  max_count = 10,
  prefer_unique = TRUE,
  prefer_binomial = TRUE,
  prefer_complete = TRUE,
  prefer_refseq = TRUE,
  only_latest = TRUE
)
}
\arguments{
\item{query}{The query to use}

\item{max_count}{The maximum number or proportion of results to download}

\item{prefer_unique}{Give preference to diverse taxa, but still return
duplicates if not enough unique taxa are found to satisfy \code{max_count}.}

\item{prefer_binomial}{Prefer genomes with a standard looking genus/species
name (no numbers)}

\item{prefer_complete}{Prefer genomes with "Complete Genome" or "Chromosome"
assembly status}

\item{prefer_refseq}{Prefer genomes in RefSeq.}

\item{only_latest}{Only include the most recent version of a genome assembly
if there are multiple versions in the result.}
}
\value{
A \code{data.frame}
}
\description{
Get metadata for NCBI assemblies associated with a query to NCBI.
}
\keyword{internal}
