% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{invalid_id_char_pattern}
\alias{invalid_id_char_pattern}
\title{Characters that cant be in IDs}
\usage{
invalid_id_char_pattern()
}
\value{
A character \code{vector} of length 1
}
\description{
Regular expression for characters that cannot appear in IDs
}
\examples{
invalid_id_char_pattern()

}
