% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperformance_aggregated.R
\name{iperformance_aggregated}
\alias{iperformance_aggregated}
\title{Construct the aggregated performance spectrum in an interactive Shiny gadget}
\usage{
iperformance_aggregated(eventlog, plotly = FALSE)
}
\arguments{
\item{eventlog}{The Event Log object on which you want to mine the performance spectrum}

\item{plotly}{True if you desire a ggplotly object as visualisation instead of a ggplot}
}
\value{
A ggplot2 or plotly object describing the aggregated performance spectrum
}
\description{
Construct the aggregated performance spectrum in an interactive Shiny gadget
}
\examples{
iperformance_aggregated(sepsis, TRUE)
}
