% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperformance_detailed.R
\name{iperformance_detailed}
\alias{iperformance_detailed}
\title{Construct the detailed performance spectrum in an interactive Shiny gadget}
\usage{
iperformance_detailed(eventlog, plotly = FALSE)
}
\arguments{
\item{eventlog}{The Event Log object on which you want to mine the performance spectrum}

\item{plotly}{True if you desire a ggplotly object as visualisation instead of a ggplot}
}
\value{
A ggplot2 or plotly object describing the detailed performance spectrum
}
\description{
Construct the detailed performance spectrum in an interactive Shiny gadget
}
\examples{
\dontrun{
iperformance_detailed(sepsis, TRUE)
}
}
