% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{known_data_types}
\alias{known_data_types}
\title{Types of data supported by the pipeline}
\usage{
known_data_types()
}
\value{
A character \code{vector}
}
\description{
These are the valid values that can be supplied to the \code{data_type} column.
}
\examples{
known_data_types()

}
