% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{known_input_columns}
\alias{known_input_columns}
\title{Columns used by \code{pathogensurveillance}}
\usage{
known_input_columns()
}
\value{
A \code{vector} of column names
}
\description{
Column in input tables that are used by the \code{pathogensurveillance} pipeline.
Other columns can be given, but only these are used.
}
\examples{
known_input_columns()

}
