% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sample_metadata.R
\name{make_ani_nj_tree}
\alias{make_ani_nj_tree}
\title{Print a table of sample metadata}
\usage{
make_ani_nj_tree(sample_data, interactive = TRUE)
}
\arguments{
\item{sample_data}{A tibble/data.frame with the sample metadata}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}
}
\description{
Selects columns in the sample metadata to print and format the result for
use as a static table in PDF or an interactive table in HTML.
}
