% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make an heatmap with dendrogram.}
\usage{
make_heatmap(
  input_matrix,
  ref_data,
  sample_data,
  interactive = FALSE,
  height = NULL,
  width = NULL,
  dpi = 100,
  font_size = 10,
  max_label_length = 30
)
}
\arguments{
\item{input_matrix}{Approximate ANI matrix output from Sourmash analysis}

\item{ref_data}{A tibble/data.frame with information on references used in analysis}

\item{sample_data}{A tibble/data.frame with the sample metadata}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}

\item{height}{The height in pixels. If not \code{interactive}, this is divided by \code{dpi} to convert it to inches.}

\item{width}{The width in pixels. If not \code{interactive}, this is divided by \code{dpi} to convert it to inches.}

\item{dpi}{How pixels are converted to inches}

\item{font_size}{Size of text used for labels}

\item{max_label_length}{Labels longer than this length will be shortened.}
}
\value{
A heatmap with dendrogram
}
\description{
The heatmap is based on a distance matrix
}
\keyword{internal}
