% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maps.R
\name{make_maps}
\alias{make_maps}
\title{Build maps}
\usage{
make_maps(input, sample_data, interactive = knitr::is_html_output())
}
\arguments{
\item{sample_data}{}
}
\value{
Maps
}
\description{
Sample maps built using Leaflet and lat/lon coordinates in input spreadsheet
}
