% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_MSNs.R
\name{make_MSN}
\alias{make_MSN}
\title{Make Minimum spanning network}
\usage{
make_MSN(
  snp_alignment_path,
  sample_data,
  population = NULL,
  interactive = FALSE,
  snp_threshold = NULL,
  show_MLG_table = FALSE,
  user_seed = NULL
)
}
\value{
minimum spanning network
}
\description{
Make Minimum spanning network
}
