% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{make_path_table_with_group}
\alias{make_path_table_with_group}
\title{Make a table with paths and report group}
\usage{
make_path_table_with_group(paths, path_func)
}
\description{
Make a table with paths and report group
}
\keyword{internal}
