% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_core_gene_phylogeny.R
\name{make_phylogeny}
\alias{make_phylogeny}
\title{Make core gene phylogeny}
\usage{
make_phylogeny(
  core_phylogeny_path,
  sample_data,
  sourmash_ani_matrix,
  reference_data,
  interactive = TRUE
)
}
\arguments{
\item{interactive}{}
}
\value{
Core gene phylogeny
}
\description{
The tree will either be made using ggtree, if output is a PDF, or phylo
}
