% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{multi_input_columns}
\alias{multi_input_columns}
\title{Columns that can have multiple values}
\usage{
multi_input_columns()
}
\value{
A character \code{vector}
}
\description{
Columns that can have multiple values separated by ';'.
}
\examples{
multi_input_columns()

}
