% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snp_data.R
\name{parse_alignment_fastas}
\alias{parse_alignment_fastas}
\title{Read SNP alignment files}
\usage{
parse_alignment_fastas(snp_align_paths, group)
}
\arguments{
\item{group}{The name of the report group this tree is for}

\item{snp_tree_paths}{One or more paths to .fasta alignment files}
}
\description{
Reads SNP alignment .fasta files produced by variant calling into R.
}
