% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_ani_matrix_names}
\alias{parse_ani_matrix_names}
\title{Parse and rename ANI matrix names for downstream steps}
\usage{
parse_ani_matrix_names(sourmash_ani_matrix)
}
\arguments{
\item{path}{the file path to 'sample_data.csv' file}

\item{group}{parsed information on report group for specific set of samples}
}
\value{
ANI matrix with column and row names that are compatible with other data inputs
}
\description{
Parse and rename ANI matrix names for downstream steps
}
