% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{parse_path_surveil_input}
\alias{parse_path_surveil_input}
\title{Parse one or more inputs to \code{pathogensurveillance}}
\usage{
parse_path_surveil_input(input_paths, fail_on_warning = FALSE)
}
\arguments{
\item{input_paths}{The paths to one or more tables in TSV, CSV, or ODS
(LibreOffice Calc), xls, or xlsx files (Excel) format.}

\item{fail_on_warning}{If \code{TRUE}, things that will normally cause a warning
will instead cause an error.}
}
\description{
Parse and validate one or more input table of sample/reference metadata to
the \code{pathogensurveillance} pipeline and puts the parsed data in a single
table.
}
\examples{

input_paths <- system.file(package = 'psminer',
                           file.path('extdata', c(
                             'feature_test.tsv',
                             'feature_test_refs.csv',
                             'ods_example.ods'
                           )))
parse_path_surveil_input(input_paths)

}
