% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_pocp_matrix}
\alias{parse_pocp_matrix}
\title{Parse POCP matrix}
\usage{
parse_pocp_matrix(path)
}
\arguments{
\item{path}{the file path to the 'popc.tsv' file}
}
\value{
POCP matrix with column and row names that are compatible with other data inputs
}
\description{
Parse and rename POCP matrix for downstream steps
}
