% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_ref_meta}
\alias{parse_ref_meta}
\title{Parse reference metadata files}
\usage{
parse_ref_meta(ref_data_path, ref_ids_path, sample_data)
}
\arguments{
\item{sample_data}{}
}
\value{
Reformatted reference metadata table containing both refseq-defined reference assemblies and user-defined assemblies
}
\description{
Parse reference metadata files
}
