% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_sample_meta}
\alias{parse_sample_meta}
\title{Parse sample metadata files}
\usage{
parse_sample_meta(path, group)
}
\arguments{
\item{group}{}
}
\description{
Parse sample metadata files
}
