% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_trees.R
\name{parse_snp_trees}
\alias{parse_snp_trees}
\title{Read SNP trees}
\usage{
parse_snp_trees(snp_tree_paths, group)
}
\arguments{
\item{snp_tree_paths}{One or more paths to .treefiles}

\item{group}{The name of the report group this tree is for}
}
\description{
Reads .treefile files produced by variant calling into R.
}
