% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_aggregated.R
\name{performance_aggregated}
\alias{performance_aggregated}
\title{Construct the aggregated performance spectrum in the plot viewer}
\usage{
performance_aggregated(
  eventlog,
  segment_coverage = NULL,
  n_segments = NULL,
  classification_attribute = NULL,
  grouping = c("start", "stop"),
  bins = NULL
)
}
\arguments{
\item{eventlog}{The Event Log object on which you want to mine the performance spectrum}

\item{segment_coverage}{The percentage of cases in which each segment must be present to be visualised in the spectrum. Dominated by \code{n_segments}}

\item{n_segments}{Visualise only the top n segments in terms of frequency}

\item{classification_attribute}{The variable defining the color legend. If \code{NULL} or \code{quartile}, a quartile variable dividing the durations of the segments in quartiles is calculated}

\item{grouping}{The timestamps, start or complete, which are binned in the histogram}

\item{bins}{The number of bins in the histogram. If left at \code{NULL}, 30 bins are selected automatically}
}
\value{
A ggplot2 object describing the aggregated performance spectrum
}
\description{
Construct the aggregated performance spectrum in the plot viewer
}
\examples{
performance_aggregated(sepsis, "quartile", "start", 15, 0.2)
}
