% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.ps_aggregated}
\alias{plot.ps_detailed}
\title{Plot Methods}
\usage{
\method{plot}{ps_aggregated}(x, ...)

\method{plot}{ps_detailed}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link[=ps_aggregated]{ps_aggregated()}} or \code{\link[=ps_detailed]{ps_detailed()}}.}

\item{...}{Additional variables to pass further.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2}} object, which can be customised further.
}
\description{
Visualize performance spectrum.
}
