% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_idtb.R
\name{plot_sendsketch_taxonomy}
\alias{plot_sendsketch_taxonomy}
\title{Make sunburst plot of sendsketch taxonomy}
\usage{
plot_sendsketch_taxonomy(sketch_data, ...)
}
\description{
Make sunburst plot of sendsketch taxonomy
}
