% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{pocp_matrix_parsed}
\alias{pocp_matrix_parsed}
\title{Get POCP matrix}
\usage{
pocp_matrix_parsed(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
a \code{list} of POCP matrices
}
\description{
Return a list of \code{\link[base:data.frame]{base::data.frame()}}s with the POCP values based on a core
gene analysis using pirate.
}
