% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_best_match_table.R
\name{print_best_match_table}
\alias{print_best_match_table}
\title{Print best match table}
\usage{
print_best_match_table(
  pairwise_matrix,
  sample_data,
  ref_data,
  metric_name,
  interactive = knitr::is_html_output(),
  ...
)
}
\arguments{
\item{pairwise_matrix}{A triangular `matrix` with all pairwise comparisons between samples and references}

\item{sample_data}{The sample metadata}

\item{ref_data}{The reference metadata}

\item{metric_name}{The name of the metric used in the matrix. This is only used to name columns in the output.}

\item{interactive}{Whether to print an interactive HTML-based table or one for use with PDF}

\item{...}{Passed to `DT::datatable`.}
}
\value{
Returns the table to print
}
\description{
Prints a table with the best matches for each sample given a pairwise distance matrix.
}
