% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_figures_with_selector.R
\name{print_figures_with_selector}
\alias{print_figures_with_selector}
\title{Print figures with dropdown selector}
\usage{
print_figures_with_selector(
  plot_func,
  selector,
  label,
  img_class,
  imglist_class = paste0(img_class, "_list"),
  ...
)
}
\arguments{
\item{plot_func}{A function to produce a plot using a single input from the \code{selector} input.}

\item{selector}{One or more values given to \code{plot_func} to make plots. There will be one plot per input}

\item{label}{Text used to label the selector dropdown.}

\item{img_class}{The CSS class used as an ID for the \code{img} HTML element. This must be unique.}

\item{imglist_class}{The CSS class used as an ID for the dropdown HTML element. This must be unique.}

\item{...}{Passed to \code{\link[grDevices:png]{grDevices::png()}}}
}
\description{
Prints HTML code to show a plot based on the value in a dropdown selector.
For use with Quarto/Rmarkdown/Knitr, put in a chunk with the option \code{results='asis'}.
}
\examples{
n <- c(10, 100, 1000)
base_plot_func <- function(x) {
  hist(rnorm(x))
}
print_figures_with_selector(base_plot_func, n, 'Count', 'base_test_id')

ggplot_func <- function(x) {
  df <- data.frame(
    sex=factor(rep(c("F", "M"), each=x)),
    weight=round(c(rnorm(x, mean=55, sd=5), rnorm(x, mean=65, sd=5)))
  )
  print(ggplot(df, aes(x=weight)) + geom_histogram())
}
print_figures_with_selector(ggplot_func, n, 'Number', 'ggplot_test_id')
}
