% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ani_table.R
\name{print_pocp_table}
\alias{print_pocp_table}
\title{Print best POCP table}
\usage{
print_pocp_table(
  pairwise_matrix,
  sample_data,
  ref_data,
  interactive = knitr::is_html_output(),
  ...
)
}
\arguments{
\item{pairwise_matrix}{A \code{matrix} with all pairwise comparisons between samples and references}

\item{sample_data}{The sample metadata}

\item{ref_data}{The reference metadata}

\item{interactive}{Whether to print an interactive HTML-based table or one for use with PDF}

\item{...}{Passed to \code{DT::datatable}.}
}
\value{
Returns the table to print
}
\description{
Prints a table with the highest  for each sample given a pairwise distance matrix.
}
