% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_detailed.R
\name{ps_detailed}
\alias{ps_detailed}
\title{Construct the detailed performance spectrum in the plot viewer}
\usage{
ps_detailed(eventlog, segment_coverage, n_segments, classification_attribute)
}
\arguments{
\item{eventlog}{The Event Log object on which you want to mine the performance spectrum}

\item{segment_coverage}{The percentage of cases in which each segment must be present in order to be visualised in the spectrum. Defaults to 0.2 (20%). Ignored if n_segments is specified.}

\item{n_segments}{Alternative to segments_coverage, Visualise only the top n segments in terms of frequency.}

\item{classification_attribute}{The variable defining the color legend. If \code{NULL} or \code{quartile}, a quartile variable dividing the durations of the segments in quartiles is calculated}
}
\value{
A ggplot2 object describing the detailed performance spectrum
}
\description{
Construct the detailed performance spectrum in the plot viewer
}
\examples{
performance_detailed(eventlog = sepsis, segments_coverage = 0.2)

}
