% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{read_input_tables}
\alias{read_input_tables}
\title{Read input tables from supported formats}
\usage{
read_input_tables(input_paths, read_all_sheets = TRUE)
}
\arguments{
\item{read_all_sheets}{If \code{TRUE}, read all sheets in spreadsheet files like
.ods and .xlsx}
}
\value{
A \code{list} of \href{tibble::tibble()}{tibble}
}
\description{
Read input tables from supported formats
}
\examples{
input_paths <- system.file(c('extdata/xls_example.xls', 'extdata/feature_test.tsv'), package = 'psminer')
psminer:::read_input_tables(input_paths)

}
\keyword{internal}
