% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_finders.R
\name{ref_meta_path}
\alias{ref_meta_path}
\title{Find reference metadata file path}
\usage{
ref_meta_path(path)
}
\arguments{
\item{path}{The path to one or more folders that contain pathogensurveillance output.}
}
\value{
character vector of length 1
}
\description{
Return the file path to the TSV containing the reference metadata for a given
pathogensurveillance output folder.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
ref_meta_path(path)

}
\seealso{
Other path finders: 
\code{\link{busco_ref_path}()},
\code{\link{busco_tree_path}()},
\code{\link{considered_ref_meta_path}()},
\code{\link{core_ref_path}()},
\code{\link{core_tree_path}()},
\code{\link{estimated_ani_matrix_path}()},
\code{\link{pocp_matrix_path}()},
\code{\link{run_info_path}()},
\code{\link{sample_meta_path}()},
\code{\link{selected_ref_meta_path}()},
\code{\link{sendsketch_path}()},
\code{\link{software_version_path}()},
\code{\link{status_message_path}()},
\code{\link{variant_align_path}()},
\code{\link{variant_ref_path}()},
\code{\link{variant_tree_path}()}
}
\concept{path finders}
