% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{report_group_path_table}
\alias{report_group_path_table}
\title{Find the report group file path data}
\usage{
report_group_path_table(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
\code{tibble} with \code{report_group_id} and \code{path} columns
}
\description{
Return a table with the file containing the name of the report group for a
given pathogensurveillance output folder.
}
\seealso{
Other path tables: 
\code{\link{busco_ref_path_table}()},
\code{\link{busco_tree_path_table}()},
\code{\link{core_ref_path_table}()},
\code{\link{ref_meta_path_table}()},
\code{\link{sample_meta_path_table}()}
}
\concept{path tables}
