% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{run_info_parsed}
\alias{run_info_parsed}
\title{Get parsed report groups}
\usage{
run_info_parsed(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \code{\link[base:character]{base::character()}} vector with the groups from all input paths
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with report group ID. The groups found in the given
paths will be combined.
}
