% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sample_meta_parsed}
\alias{sample_meta_parsed}
\title{Get parsed sample metadata}
\usage{
sample_meta_parsed(path)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the sample metadata
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with sample metadata. The contents of all
sample metadata files found in the given paths will be combined.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
sample_meta_parsed(path)

}
