% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{sample_meta_path_data}
\alias{sample_meta_path_data}
\title{Find sample metadata path data}
\usage{
sample_meta_path_data(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain pathogensurveillance output.}
}
\value{
\code{tibble} with \code{report_group_id} and \code{path} columns
}
\description{
Return the file path data to the TSV containing the sample metadata for a given
pathogensurveillance output folder.
}
\seealso{
Other path tables: 
\code{\link{busco_ref_path_data}()},
\code{\link{busco_tree_path_data}()},
\code{\link{considered_ref_meta_path_data}()},
\code{\link{core_ref_path_data}()},
\code{\link{core_tree_path_data}()},
\code{\link{estimated_ani_matrix_path_data}()},
\code{\link{pocp_matrix_path_data}()},
\code{\link{ref_meta_path_data}()},
\code{\link{run_info_path_data}()},
\code{\link{selected_ref_meta_path_data}()},
\code{\link{sendsketch_path_data}()},
\code{\link{software_version_path_data}()},
\code{\link{status_message_path_data}()},
\code{\link{variant_align_path_data}()},
\code{\link{variant_ref_path_data}()},
\code{\link{variant_tree_path_data}()}
}
\concept{path tables}
