% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{sample_meta_table}
\alias{sample_meta_table}
\title{Print a table of sample metadata}
\usage{
sample_meta_table(input, interactive = FALSE, ...)
}
\arguments{
\item{input}{The path to one or more folders that contain
pathogensurveillance output or a table in the format of the
\code{\link[=sample_meta_parsed]{sample_meta_parsed()}} output.}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}

\item{...}{Passed to \code{DT::datatable}.}
}
\description{
Selects columns in the sample metadata to print and format the result for
use as a static table in PDF or an interactive table in HTML.
}
