% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendsketch.R
\name{sendsketch_best_hits}
\alias{sendsketch_best_hits}
\title{Pick best sendsketch hits for each sample}
\usage{
sendsketch_best_hits(
  sketch_data,
  sort_columns = c("WKID", "ANI", "Complt"),
  top_n = 1
)
}
\arguments{
\item{sketch_data}{A dataframe containing sketch analysis results.}

\item{sort_columns}{Character vector; specifies the columns to sort the table by in descending order.
The default is \code{c("sample_id", "WKID", "ANI", "Complt")}. The \code{sample_id} is expected to be
unique for each row, and the table will be sorted by this column first, followed by the other
metrics in the order provided.}

\item{top_n}{Integer; the number of top entries to retain for each unique sample ID after sorting.
The default is \code{1}, meaning only the top entry is kept.}
}
\value{
A \code{data.frame}
}
\description{
This function processes sendsketch data from pathogen surveillance nextflow pipeline and
filters the output so that only the best hits are present
}
\examples{
# Assuming `sketch_data` is your dataframe with the appropriate structure:
sendsketch_best_hits(sketch_data, sort_columns = c("sample_id", "WKID", "ANI", "Complt"), top_n = 1)

# If you want to sort by Completeness and then WKID, keeping the top 2 entries for each sample_id:
sendsketch_best_hits(sketch_data, sort_columns = c("sample_id", "Complt", "WKID"), top_n = 2)
}
