% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_parsed}
\alias{sendsketch_parsed}
\title{Get parsed sendsketch results}
\usage{
sendsketch_parsed(paths, only_best = FALSE)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}

\item{only_best}{Only return the best hit for each combination of report
group and sample. For more control/details on how the top hit is selected,
see \code{\link[=sendsketch_best_hits]{sendsketch_best_hits()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the sendsketch output combined
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with the results from sendsketch. The
contents of all sendsketch files found in the given paths will be combined.
}
