% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_taxonomy_data_parsed}
\alias{sendsketch_taxonomy_data_parsed}
\title{Get parsed sendsketch taxonomy data}
\usage{
sendsketch_taxonomy_data_parsed(path, only_best = FALSE, only_shared = FALSE)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{only_best}{Only return the best hit for each combination of report
group and sample. For more control/details on how the top hit is selected,
see \code{\link[=sendsketch_best_hits]{sendsketch_best_hits()}}.}

\item{only_shared}{If \code{TRUE}, only return the data for that are present in
all of the inputs.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with taxonomy data, with columns corresponding
to ranks.
}
\description{
Return the taxonomic data in sendsketch results associated with directory
paths containing \code{pathogensurveillance} output.
}
\examples{
path <- system.file('extdata/ps_output', package = 'psminer')
sendsketch_taxonomy_data_parsed(path)
sendsketch_taxonomy_data_parsed(path, only_best = TRUE)
sendsketch_taxonomy_data_parsed(path, only_shared = TRUE)

}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{multigene_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
