% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_taxonomy_parsed}
\alias{sendsketch_taxonomy_parsed}
\title{Get parsed sendsketch taxonomy}
\usage{
sendsketch_taxonomy_parsed(paths, remove_ranks = FALSE, only_best = FALSE)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}

\item{remove_ranks}{If \code{TRUE}, remove the rank information from the taxonomy.}

\item{only_best}{Only return the best hit for each combination of report
group and sample. For more control/details on how the top hit is selected,
see \code{\link[=sendsketch_best_hits]{sendsketch_best_hits()}}.}
}
\value{
A \code{\link[base:character]{base::character()}} vector of taxonomic classifications, each
delimited with \verb{;}, named by sample IDs.
}
\description{
Return the taxonomic classificaiton in sendsketch results associated with
directory paths containing \code{pathogensurveillance} output.
}
