% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{shared_char}
\alias{shared_char}
\title{Identify conserved start/ends of characters}
\usage{
shared_char(col, end = FALSE)
}
\arguments{
\item{col}{The character vector}

\item{end}{If `TRUE` then return shared end instead of shared start}
}
\description{
Find any starts or ends of strings that are the same in all elements of a character vector
}
\keyword{internal}
