% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_idtb.R
\name{sketch_idtb}
\alias{sketch_idtb}
\title{Create an summary table from sendsketch output table}
\usage{
sketch_idtb(
  sketch_data,
  sort_columns = c("sample_id", "WKID", "ANI", "Complt"),
  top_n = 1
)
}
\arguments{
\item{sketch_data}{A dataframe containing sketch analysis results with the following columns:}

\item{sort_columns}{Character vector; specifies the columns to sort the table by in descending order.
The default is `c("sample_id", "WKID", "ANI", "Complt")`. The `sample_id` is expected to be
unique for each row, and the table will be sorted by this column first, followed by the other
metrics in the order provided.}

\item{top_n}{Integer; the number of top entries to retain for each unique sample ID after sorting.
The default is `1`, meaning only the top entry is kept.}

\item{sample_id}{Character; unique identifiers for each sample.}

\item{WKID}{Numeric; Weighted Kmer IDentity percentages indicating the kmer-based similarity.}

\item{ANI}{Numeric; Average Nucleotide Identity percentages representing genomic similarity.}

\item{Complt}{Numeric; Completeness percentages reflecting the coverage of the genome.}

\item{taxName}{Character; taxonomic names providing the preliminary identification of the sample.}
}
\value{
An interactive DT::datatable object that renders as an HTML table when used in an R Markdown document or Shiny application.
The table will have interactive features such as sorting and search enabled.
}
\description{
This function processes sendsketch data from pathogen surveillance nextflow pipeline and
generates an interactive HTML table displaying the top hits from the sketch-based analysis.
The resulting table includes metrics such as Weighted Kmer IDentity (WKID),
Average Nucleotide Identity (ANI), and Completeness, providing insights into
genomic similarity and representational completeness between the query and reference genomes.
The table can be sorted by any of these metrics, and only the top results as specified
by `top_n` are displayed for each unique sample ID. Additionally, the table
visually encodes the percentage values using horizontal bars for an intuitive
and accessible presentation.
}
\examples{
# Assuming `sketch_data` is your dataframe with the appropriate structure:
sketch_idtb(sketch_data, sort_columns = c("sample_id", "WKID", "ANI", "Complt"), top_n = 1)

# If you want to sort by Completeness and then WKID, keeping the top 2 entries for each sample_id:
sketch_idtb(sketch_data, sort_columns = c("sample_id", "Complt", "WKID"), top_n = 2)
}
