% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{software_version_parsed}
\alias{software_version_parsed}
\title{Parse Software Version Metadata from YAML File}
\usage{
software_version_parsed(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{module}, \code{program}, and \code{version} detailing the software versions.
}
\description{
Reads a YAML file containing software version information and transforms it into a tibble.
Each software module and program, along with its version, is extracted and organized.
}
\examples{
version_path <- file.path(params$inputs, "inputs", "versions.yml")
version_data <- parse_software_meta(version_path)
}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{multigene_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
