% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{status_message_parsed}
\alias{status_message_parsed}
\title{Get parsed pipeline status data}
\usage{
status_message_parsed(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the messages from all input paths
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with the status messages produced by the pathogensuriveillance
pipeline. The contents of all status message files found in the given paths
will be combined.
}
