% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{status_message_parsed_summary}
\alias{status_message_parsed_summary}
\title{Get summary of pipeline status data}
\usage{
status_message_parsed_summary(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with counts of messages attributes
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with the numbers of issues encountered by the
pipeline. The contents of all status message files found in the given paths
will be combined.
}
\examples{
path <- system.file('extdata/ps_output', package = 'psminer')
status_message_parsed_summary(path)

}
