% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{status_message_table}
\alias{status_message_table}
\title{Get table of pipeline status data}
\usage{
status_message_table(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \link[DT:datatable]{DT::datatable} with details for errors, warnings, notes.
}
\description{
Return a formatted interactive table with the data on the issues encountered
by the pipeline, one row for each issue. The contents of all status message
files found in the given paths will be combined.
}
