% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{status_message_table_summary}
\alias{status_message_table_summary}
\title{Get summary table of pipeline status data}
\usage{
status_message_table_summary(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
A \link[DT:datatable]{DT::datatable} with counts of errors, warnings, notes.
}
\description{
Return a formatted interactive table with the numbers of issues encountered
by the pipeline, one row for each pipeline step. The contents of all status
message files found in the given paths will be combined.
}
