% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tree_parsed}
\alias{tree_parsed}
\title{Get phylogenies using a function}
\usage{
tree_parsed(paths, func)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
List of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path
}
\description{
Return a list of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path by searching for
folders containing pathogensurveillance output.
}
\keyword{internal}
