% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{valid_data_types}
\alias{valid_data_types}
\title{Values for the \code{data_type} column}
\usage{
valid_data_types()
}
\value{
A character \code{vector}
}
\description{
Valid values for the \code{data_type} column of \code{pathogensurveillance} input
tables.
}
\examples{
valid_data_types()

}
