% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{valid_input_type}
\alias{valid_input_type}
\title{Values for the \code{input_type} column}
\usage{
valid_input_type()
}
\value{
A character \code{vector}
}
\description{
Valid values for the \code{input_type} column of \code{pathogensurveillance} input
tables.
}
\examples{
valid_input_type()

}
