% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{valid_ref_usages}
\alias{valid_ref_usages}
\title{Values for the \verb{ref_*_usage} columns}
\usage{
valid_ref_usages()
}
\value{
A character \code{vector}
}
\description{
Valid values for the \code{ref_primary_usage} and \code{ref_contextual_usage} columns
of \code{pathogensurveillance} input tables.
}
\examples{
valid_ref_usages()

}
