% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{valid_types}
\alias{valid_types}
\title{Values for the \code{type} column}
\usage{
valid_types()
}
\value{
A character \code{vector}
}
\description{
Valid values for the \code{type} column of \code{pathogensurveillance} input
tables.
}
\examples{
valid_types()

}
