% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{valid_usages}
\alias{valid_usages}
\title{Values for the \code{usage} column}
\usage{
valid_usages()
}
\value{
A character \code{vector}
}
\description{
Valid values for the \code{usage} column of \code{pathogensurveillance} input
tables.
}
\examples{
valid_usages()

}
