% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{validate_data_source_paths}
\alias{validate_data_source_paths}
\title{Check value of \code{data_source} for \code{data_type}s that contain paths}
\usage{
validate_data_source_paths(metadata, data_type, max_paths, valid_extensions)
}
\arguments{
\item{metadata}{the input table}

\item{data_type}{the value of the \code{data_type} column to validate}

\item{max_paths}{The maximum number of paths that can be present. Set to
\code{NULL} to disable this check.}

\item{valid_extensions}{A \code{vector} of valid extensions. Set to \code{NULL} to
disable this check.}
}
\description{
Check value of \code{data_source} for \code{data_type}s that contain paths
}
\keyword{internal}
