% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{validate_source_generic}
\alias{validate_source_generic}
\title{Check that \code{source} counts and values}
\usage{
validate_source_generic(
  metadata,
  type,
  must_exist = TRUE,
  min_count = 1,
  max_count = NULL,
  starts_with = NULL,
  ends_with = NULL,
  default = NULL
)
}
\arguments{
\item{metadata}{the input table}

\item{type}{the value of the \code{type} column to validate}

\item{must_exist}{If \code{TRUE}, it cannot be \code{NA} or blank. This check does not
consider the presence of \verb{;} so it is safe for values where \verb{;} is not used
a separater.}

\item{min_count}{The minimum number of values that can be present. Set to
\code{NULL} to disable this check.}

\item{max_count}{The maximum number of values that can be present. Set to
\code{NULL} to disable this check.}

\item{starts_with}{A \code{vector} of values that a valid value can have as a
prefix. Set to \code{NULL} to disable this check.}

\item{ends_with}{A \code{vector} of values that a valid value can have as a
suffix. Set to \code{NULL} to disable this check.}

\item{default}{What is put if a value is missing.}
}
\description{
Check that \code{source} has the right number of values and starts/ends with
the right values.
}
\keyword{internal}
