% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{variant_align_parsed}
\alias{variant_align_parsed}
\title{Find and parse SNP alignments}
\usage{
variant_align_parsed(path, rename = TRUE)
}
\arguments{
\item{rename}{If \code{TRUE}, rename the sequence labels as sample IDs and reference IDs.}

\item{paths}{The path to one or more folders that contain pathogensurveillance output.}
}
\value{
list of \code{\link[ape:DNAbin]{ape::DNAbin()}}, named by alignment file
}
\description{
Returns parsed SNP alignments for each reference used in the
variant analysis for a given pathogensurveillance output folder.
}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
