% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_data_finders.R
\name{variant_align_path_table}
\alias{variant_align_path_table}
\title{Find the SNP alignment path data}
\usage{
variant_align_path_table(paths)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
\code{tibble} with \code{report_group_id} and \code{path} columns
}
\description{
Return a table with the file paths of the SNP alignments for each reference
used in the variant analysis for a given pathogensurveillance output folder.
}
\seealso{
Other path tables: 
\code{\link{busco_ref_path_table}()},
\code{\link{busco_tree_path_table}()},
\code{\link{considered_ref_meta_path_table}()},
\code{\link{core_ref_path_table}()},
\code{\link{core_tree_path_table}()},
\code{\link{estimated_ani_matrix_path_table}()},
\code{\link{pocp_matrix_path_table}()},
\code{\link{ref_meta_path_table}()},
\code{\link{report_group_path_table}()},
\code{\link{sample_meta_path_table}()},
\code{\link{selected_ref_meta_path_table}()},
\code{\link{sendsketch_path_table}()},
\code{\link{software_version_path_table}()},
\code{\link{status_message_path_table}()},
\code{\link{variant_ref_path_table}()},
\code{\link{variant_tree_path_table}()}
}
\concept{path tables}
