% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{variant_msn_plot}
\alias{variant_msn_plot}
\title{Plot MSN of variant data}
\usage{
variant_msn_plot(path, combine = TRUE)
}
\arguments{
\item{combine}{If \code{TRUE} combine multiple MSNs into a single figure and
return a single figure. If \code{FALSE}, return a list of figures.}

\item{input}{The path to one or more folders that contain
pathogensurveillance output.}
}
\description{
Plot a minimum spanning network for each group of samples aligned to a
reference found in \code{pathogensurveillance} output.
}
