% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{core_tree_parsed}
\alias{core_tree_parsed}
\title{Get core gene phylogeny}
\usage{
core_tree_parsed(path)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
List of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path
}
\description{
Return a list of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path by searching for
folders containing pathogensurveillance output.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
core_tree_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{multigene_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
