% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{get_ncbi_assemblies}
\alias{get_ncbi_assemblies}
\title{Get metadata for NCBI assemblies using a query}
\usage{
get_ncbi_assemblies(
  query,
  max_count = 10,
  prefer_unique = TRUE,
  prefer_binomial = TRUE
)
}
\arguments{
\item{query}{The query to use}

\item{max_count}{The maximum number or proportion of results to download}

\item{prefer_unique}{Give preference to diverse taxa, but still return
duplicates if not enough unique taxa are found to satisfy \code{max_count}.}

\item{prefer_bionomial}{}
}
\description{
Get metadata for NCBI assemblies associated with a query to NCBI.
}
\keyword{internal}
