% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ANI_outputs.R
\name{make_ani_heatmap}
\alias{make_ani_heatmap}
\title{Make an ANI heatmap and dendrogram.}
\usage{
make_ani_heatmap(
  ani_matrix,
  ref_data,
  sample_data,
  interactive = FALSE,
  height = 1000,
  width = 1000,
  dpi = 100
)
}
\arguments{
\item{ani_matrix}{Approximate ANI matrix output from Sourmash analysis}

\item{ref_data}{A tibble/data.frame with information on references used in analysis}

\item{sample_data}{A tibble/data.frame with the sample metadata}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}

\item{height}{The height in pixels. If not \code{interactive}, this is divided by \code{dpi} to convert it to inches.}

\item{width}{The width in pixels. If not \code{interactive}, this is divided by \code{dpi} to convert it to inches.}

\item{dpi}{How pixels are converted to inches}
}
\value{
A heatmap and dendrogram
}
\description{
The ANI heatmap is based on approximate ANI similarity matrix output by Sourmash
}
\examples{
make_ani_heatmap(ani_matrix, ref_data, samp_data, interactive=FALSE)
}
