% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maps.R
\name{make_maps}
\alias{make_maps}
\title{Build maps}
\usage{
make_maps(sample_data)
}
\arguments{
\item{sample_data}{sample_data}
}
\value{
Maps
}
\description{
Sample maps built using Leaflet and lat/lon coordinates in input spreadsheet
}
