% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MSNs.R
\name{make_MSN}
\alias{make_MSN}
\title{Make Minimum spanning network}
\usage{
make_MSN(
  snp_fasta_alignment,
  sample_data,
  population = NULL,
  interactive = knitr::is_html_output(),
  snp_threshold = NULL,
  show_MLG_table = FALSE,
  user_seed = NULL,
  ...
)
}
\value{
minimum spanning network
}
\description{
Make Minimum spanning network
}
