% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{multigene_tree_parsed}
\alias{multigene_tree_parsed}
\title{Get mulitgene phylogenies}
\usage{
multigene_tree_parsed(path)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}
}
\value{
List of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path
}
\description{
Return a list of \code{\link[ape:read.tree]{ape::phylo()}} objects named by file path by searching for
folders containing pathogensurveillance output. This includes all types of
multigene trees produced by pathogensurveillance, such as the core gene
phylogeny and the busco gene phylogeny.
}
\examples{
path <- system.file('extdata/ps_output', package = 'psminer')
multigene_tree_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{sendsketch_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
