% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{parse_path_surveil_input}
\alias{parse_path_surveil_input}
\title{Parse one or more inputs to \code{pathogensurveillance}}
\usage{
parse_path_surveil_input(input_paths)
}
\arguments{
\item{input_paths}{The paths to one or more tables in TSV, CSV, or ODS
(LibreOffice Calc), xls, or xlsx files (Excel) format.}
}
\description{
Parse and validate one or more input table of sample/reference metadata to
the \code{pathogensurveillance} pipeline and puts the parsed data in a single
table.
}
\examples{

input_paths <- c(
  system.file('extdata/feature_test.tsv',      package = 'psminer'),
  system.file('extdata/feature_test_refs.csv', package = 'psminer')
)

}
