% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_ref_meta}
\alias{parse_ref_meta}
\title{Parse reference metadata files}
\usage{
parse_ref_meta(ref_data_path, assigned_refs_path, sample_data_path, group)
}
\arguments{
\item{ref_data_path}{The folder with files with the data on references
assigned to each sample. The individual files are named by sample ID.}

\item{assigned_refs_path}{The file path to the 'assigned_refs.csv' file}

\item{sample_data_path}{The file path to the 'sample_data.csv' file}

\item{group}{The name of the report group for specific set of samples.}
}
\value{
Reformatted reference metadata table containing both refseq-defined
reference assemblies and user-defined assemblies
}
\description{
Parse reference metadata files
}
