% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_sample_meta}
\alias{parse_sample_meta}
\title{Parse sample metadata files}
\usage{
parse_sample_meta(path, group)
}
\arguments{
\item{path}{the file path to 'samp_data.csv' file}

\item{group}{parsed information on report group for specific set of samples}
}
\value{
samp_data dataframe
}
\description{
Parse sample metadata files
}
