% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_tables.R
\name{parse_sample_meta}
\alias{parse_sample_meta}
\title{Parse sample metadata files}
\usage{
parse_sample_meta(sample_data_path, assigned_refs_path, group)
}
\arguments{
\item{sample_data_path}{The file path to the 'sample_data.tsv' file}

\item{assigned_refs_path}{The file path to the 'assigned_refs.tsv' file}

\item{group}{The name of the report group for specific set of samples.}
}
\value{
\code{data.frame} of metadata for samples, with one row for each sample in
the group
}
\description{
Parse sample metadata files
}
