% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.ps_aggregated}
\title{Plot Methods}
\usage{
\method{plot}{ps_aggregated}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ps_aggregated} or \code{ps_detailed}.}

\item{...}{Additional variables to pass further.}
}
\value{
A \code{\link{ggplot2}} object, which can be customised further.
}
\description{
Visualize metric
}
