% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Plot a phylogeny}
\usage{
plot_phylogeny(
  trees,
  sample_meta,
  ref_meta,
  color_by = NULL,
  collapse_by_tax = NULL,
  interactive = FALSE,
  tip_label_size = NULL,
  node_label_size = NULL,
  ...
)
}
\arguments{
\item{trees}{One or more \code{\link[ape:read.tree]{ape::phylo()}} objects. Inputting multiple trees
only makes sense when \code{collapse_by_tax} is used.}

\item{sample_meta}{A table containing metadata for samples in \code{tree_paths}.}

\item{ref_meta}{A table containing metadata for references in \code{tree_paths}.}

\item{color_by}{The name of the column in the metadata used to color samples.}

\item{collapse_by_tax}{A \code{\link[tibble:tibble]{tibble::tibble()}} with the columns \code{sample_id} and
those named by taxonomic ranks. These are ids present in \code{sample_meta}.
These are used to provide a taxo`nomic tree that functions as a backbone to
combine multiple trees into a single one. (Default: return a list of plots
instead of a single plot)}

\item{interactive}{Whether to use an HTML-based interactive format or not
(default: TRUE)}

\item{tip_label_size}{The size of tree tip labels. 1-5 might be reasonable values.}

\item{node_label_size}{The size of tree node (bootstrap) labels. 1-5 might be reasonable values.}
}
\description{
Plot a phylogeny
}
\keyword{internal}
