% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_core_gene_phylogeny.R, R/make_trees.R
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Plot a phylogeny}
\usage{
plot_phylogeny(
  phylo_path,
  ids,
  labels = ids,
  colors = "#000000",
  interactive = knitr::is_html_output(),
  ...
)

plot_phylogeny(
  phylo_path,
  ids,
  labels = ids,
  colors = "#000000",
  interactive = knitr::is_html_output(),
  ...
)
}
\arguments{
\item{phylo_path}{Path to the .treefile}

\item{ids}{The ids used to identify tips in \code{phylo_path}. The tree will be subset to these tips.}

\item{labels}{The labels to print on the plot. Must be the same length as \code{ids}.}

\item{colors}{The color or the tips. Must be the same length as \code{ids}.}

\item{interactive}{Whether to use an HTML-based interactive format or not (default: TRUE)}
}
\description{
Plot a phylogeny

Plot a phylogeny
}
