% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendsketch.R
\name{plot_sendsketch_taxonomy}
\alias{plot_sendsketch_taxonomy}
\title{Make sunburst plot of sendsketch taxonomy}
\usage{
plot_sendsketch_taxonomy(
  sketch_data,
  interactive = knitr::is_html_output(),
  ...
)
}
\arguments{
\item{sketch_data}{A dataframe containing sketch analysis results.}

\item{interactive}{Whether or not to produce an interactive HTML/javascript-based plot or a static one.}
}
\description{
Converts classifications of top hits in sendsketch output into an interactive sunburst plot.
}
