% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_aggregated.R
\name{ps_aggregated}
\alias{ps_aggregated}
\alias{ps_aggregated.log}
\title{Aggregated Performance Spectrum}
\usage{
ps_aggregated(
  log,
  segment_coverage = 0.2,
  n_segments,
  classification_attribute = NULL,
  grouping = c("start", "stop"),
  bins = NULL
)

\method{ps_aggregated}{log}(
  log,
  segment_coverage = 0.2,
  n_segments,
  classification_attribute = NULL,
  grouping = c("start", "stop"),
  bins = NULL
)
}
\arguments{
\item{log}{\code{\link[bupaR]{log}}: Object of class \code{\link[bupaR]{log}} or derivatives (\code{\link[bupaR]{grouped_log}},
\code{\link[bupaR]{eventlog}}, \code{\link[bupaR]{activitylog}}, etc.).}

\item{segment_coverage, n_segments}{\code{\link{numeric}}: Provide either \code{segment_coverage} or \code{n_segments}. If neither is
provided, \code{segment_coverage = 0.2} will be used.\cr
\code{segment_coverage}: The percentage of cases (default \code{0.2}) in which each segment must be present to be visualised in the spectrum.
Ignored if \code{n_segments} is specified.\cr
\code{n_segments}: Visualise only the top \code{n} segments based on frequency.}

\item{classification_attribute}{\code{\link{character}} (default \code{\link{NULL}}): The variable defining the colour legend.
If \code{\link{NULL}} or \code{"quartile"}, a quartile variable dividing the durations of the segments in quartiles is calculated.}

\item{grouping}{\code{\link{character}} (default \code{"start"}): The timestamps, \code{"start"} or \code{"complete"}, which are binned in the histogram.}

\item{bins}{The number of bins in the histogram. If left at \code{NULL}, 30 bins are selected automatically}
}
\value{
A \code{\link{ggplot2}} object describing the aggregated performance spectrum.
}
\description{
Construct the aggregated performance spectrum in the plot viewer.
}
\section{Methods (by class)}{
\itemize{
\item \code{ps_aggregated(log)}: Calculate aggregated performance spectrum for a \code{\link[bupaR]{log}}.

}}
\examples{
ps_aggregated(sepsis, "quartile", "start", 15, 0.2)

}
