% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsers.R
\name{read_input_tables}
\alias{read_input_tables}
\title{Read input tables from supported formats}
\usage{
read_input_tables(
  input_paths,
  read_all_sheets = TRUE,
  add_input_metadata = TRUE
)
}
\arguments{
\item{input_paths}{The paths to one or more files containing tables}

\item{read_all_sheets}{If \code{TRUE}, read all sheets in spreadsheet files like
.ods and .xlsx}

\item{add_input_metadata}{Whether to add \verb{_input_path_}, \verb{_sheet_name_} and
\verb{_row_index_} columns to each table to record where each row came from.}
}
\value{
A \code{list} of \code{tibble}.
}
\description{
When a spreadsheet format such as .ods or .xlsx is supplied, each sheet is
returned as its own table.
}
\keyword{internal}
