% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{read_input_tables}
\alias{read_input_tables}
\title{Read input tables from supported formats}
\usage{
read_input_tables(input_paths, read_all_sheets = TRUE)
}
\arguments{
\item{input_paths}{The paths to one or more files containing tables}

\item{read_all_sheets}{If \code{TRUE}, read all sheets in spreadsheet files like
.ods and .xlsx}
}
\value{
A \code{list} of \code{tibble}.
}
\description{
When a spreadsheet format such as .ods or .xlsx is supplied, each sheet is
returned as its own table.
}
\examples{
input_paths <- system.file(c('extdata/xls_example.xls', 'extdata/feature_test.tsv'), package = 'psminer')
psminer:::read_input_tables(input_paths)

}
\keyword{internal}
