% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sendsketch_parsed}
\alias{sendsketch_parsed}
\title{Get parsed sendsketch results}
\usage{
sendsketch_parsed(path, only_best = FALSE)
}
\arguments{
\item{path}{The path to one or more folders that contain
pathogensurveillance output.}

\item{only_best}{Only return the best hit for each combination of report
group and sample.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the sendsketch output combined
}
\description{
Return a \code{\link[tibble:tibble]{tibble::tibble()}} (table) with the results from sendsketch. The
contents of all sendsketch files found in the given paths will be combined.
}
\examples{
path <- system.file('extdata/ps_output', package = 'PathoSurveilR')
sendsketch_parsed(path)

}
\seealso{
Other parsers: 
\code{\link{busco_tree_parsed}()},
\code{\link{core_tree_parsed}()},
\code{\link{multigene_tree_parsed}()},
\code{\link{sendsketch_taxonomy_data_parsed}()},
\code{\link{sendsketch_taxonomy_parsed}()},
\code{\link{software_version_parsed}()},
\code{\link{tree_parsed}()},
\code{\link{variant_align_parsed}()},
\code{\link{variant_tree_parsed}()}
}
\concept{parsers}
