% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendsketch.R
\name{sendsketch_table}
\alias{sendsketch_table}
\title{Create an summary table from sendsketch output table}
\usage{
sendsketch_table(input, interactive = FALSE)
}
\arguments{
\item{input}{The path to one or more folders that contain
pathogensurveillance output or a table in the format of the
\code{\link[=sendsketch_parsed]{sendsketch_parsed()}} output. Only the best hits are returned, based on the
default behavior of \code{\link[=sendsketch_best_hits]{sendsketch_best_hits()}}. To change this behavior, pass
in the results of running \code{\link[=sendsketch_best_hits]{sendsketch_best_hits()}}.}

\item{interactive}{Whether or not to produce an interactive
HTML/javascript-based table or a static one.}
}
\value{
An interactive DT::datatable object that renders as an HTML table
when used in an R Markdown document or Shiny application. The table will
have interactive features such as sorting and search enabled.
}
\description{
This function processes sendsketch data from pathogen surveillance nextflow
pipeline and generates an interactive HTML table displaying the top hits from
the sketch-based analysis. The resulting table includes metrics such as
Weighted Kmer IDentity (WKID), Average Nucleotide Identity (ANI), and
Completeness, providing insights into genomic similarity and representational
completeness between the query and reference genomes. The table visually
encodes the percentage values using horizontal bars for an intuitive and
accessible presentation.
}
\examples{
# Assuming `sketch_data` is your dataframe with the appropriate structure:
sketch_idtb(sketch_data, sort_columns = c("sample_id", "WKID", "ANI", "Complt"), top_n = 1)

# If you want to sort by Completeness and then WKID, keeping the top 2 entries for each sample_id:
sketch_idtb(sketch_data, sort_columns = c("sample_id", "Complt", "WKID"), top_n = 2)
}
