% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_printers.R
\name{status_message_table}
\alias{status_message_table}
\title{Get table of pipeline status data}
\usage{
status_message_table(paths, interactive = knitr::is_html_output(), ...)
}
\arguments{
\item{paths}{The path to one or more folders that contain
pathogensurveillance output.}

\item{interactive}{Whether to produce interactive tables
(TRUE) or static tables (FALSE). Defaults to TRUE if the environment
supports HTML output, otherwise FALSE. Interactive tables offer enhanced
browsing capabilities, while static tables are best for printed pdf
reports.}

\item{...}{Passed to \code{DT::datatable}.}
}
\value{
A table with details for errors, warnings, notes.
}
\description{
Return a formatted interactive table with the data on the issues encountered
by the pipeline, one row for each issue. The contents of all status message
files found in the given paths will be combined.
}
